+++ packages/web3-core-method/src/index.js
@@ -56,6 +56,7 @@ var Method = function Method(options) {
     this.transactionBlockTimeout = options.transactionBlockTimeout || 50;
     this.transactionConfirmationBlocks = options.transactionConfirmationBlocks || 24;
     this.transactionPollingTimeout = options.transactionPollingTimeout || 750;
+    this.blockHeaderTimeout = options.blockHeaderTimeout || 10; // 10 seconds
     this.defaultCommon = options.defaultCommon;
     this.defaultChain = options.defaultChain;
     this.defaultHardfork = options.defaultHardfork;
@@ -204,6 +205,7 @@ Method.prototype._confirmTransaction = function (defer, result, payload) {
         timeoutCount = 0,
         confirmationCount = 0,
         intervalId = null,
+        blockHeaderTimeoutId = null,
         lastBlock = null,
         receiptJSON = '',
         gasProvided = ((!!payload.params[0] && typeof payload.params[0] === 'object') && payload.params[0].gas) ? payload.params[0].gas : null,
@@ -272,6 +274,7 @@ Method.prototype._confirmTransaction = function (defer, result, payload) {
                 sub = {
                     unsubscribe: function () {
                         clearInterval(intervalId);
+                        clearTimeout(blockHeaderTimeoutId);
                     }
                 };
             }
@@ -547,22 +550,35 @@ Method.prototype._confirmTransaction = function (defer, result, payload) {
 
     // start watching for confirmation depending on the support features of the provider
     var startWatching = function (existingReceipt) {
+        let blockHeaderArrived = false; 
+
         const startInterval = () => {
             intervalId = setInterval(checkConfirmation.bind(null, existingReceipt, true), 1000);
         };
 
-        if (!this.requestManager.provider.on) {
-            startInterval();
-        } else {
-            _ethereumCall.subscribe('newBlockHeaders', function (err, blockHeader, sub) {
-                if (err || !blockHeader) {
-                    // fall back to polling
-                    startInterval();
-                } else {
-                    checkConfirmation(existingReceipt, false, err, blockHeader, sub);
-                }
-            });
+        // If provider do not support event subscription use polling
+        if(!this.requestManager.provider.on) {
+            return startInterval();            
         }
+
+        // Subscribe to new block headers to look for tx receipt
+        _ethereumCall.subscribe('newBlockHeaders', function (err, blockHeader, sub) {
+            blockHeaderArrived = true; 
+
+            if (err || !blockHeader) {
+                // fall back to polling
+                return  startInterval();
+            }
+
+            checkConfirmation(existingReceipt, false, err, blockHeader, sub);
+        });
+
+        // Fallback to polling if tx receipt didn't arrived in "blockHeaderTimeout" [10 seconds]
+        blockHeaderTimeoutId = setTimeout(() => {
+            if(!blockHeaderArrived) {
+                startInterval();
+            }
+        }, this.blockHeaderTimeout * 1000);
     }.bind(this);
 
 
@@ -850,7 +866,10 @@ function _handleTxPricing(method, tx) {
                 getGasPrice()
             ]).then(responses => {
                 const [block, gasPrice] = responses;
-                if (block && block.baseFeePerGas) {
+                if (
+                    (tx.type === '0x2' || tx.type === undefined) &&
+                    (block && block.baseFeePerGas)
+                ) {
                     // The network supports EIP-1559
 
                     // Taken from https://github.com/ethers-io/ethers.js/blob/ba6854bdd5a912fe873d5da494cb5c62c190adde/packages/abstract-provider/src.ts/index.ts#L230